@@ -73,19 +73,19 @@
         "jest-environment-node": "^29.7.0",
         "jest-marbles": "^3.1.0",
         "jest-mock-extended": "^4.0.0-beta1",
-        "jest-preset-angular": "~14.1.0",
-        "nx": "19.6.0",
+        "jest-preset-angular": "14.4.2",
+        "nx": "20.2.2",
         "postcss": "^8.4.5",
         "prettier": "^3.2.5",
         "pretty-quick": "^4.0.0",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-refresh": "^0.10.0",
-        "storybook": "^8.2.8",
+        "storybook": "^8.4.7",
         "tailwindcss": "^3.0.2",
         "ts-jest": "^29.1.0",
         "ts-node": "10.9.1",
-        "typescript": "~5.5.2",
+        "typescript": "5.6.3",
         "vite": "^5.0.0",
         "vite-plugin-dts": "~3.8.1",
         "vitest": "^1.3.1",
@@ -93,28 +93,26 @@
         "yargs": "^17.7.2"
     },
     "dependencies": {
-        "@angular/animations": "~18.2.0",
-        "@angular/cdk": "^18.2.13",
-        "@angular/common": "~18.2.0",
-        "@angular/compiler": "~18.2.0",
-        "@angular/core": "~18.2.0",
-        "@angular/forms": "~18.2.0",
-        "@angular/platform-browser": "~18.2.0",
-        "@angular/platform-browser-dynamic": "~18.2.0",
-        "@angular/router": "~18.2.0",
+        "@angular/animations": "19.0.5",
+        "@angular/cdk": "19.0.4",
+        "@angular/common": "19.0.5",
+        "@angular/compiler": "19.0.5",
+        "@angular/core": "19.0.5",
+        "@angular/forms": "19.0.5",
+        "@angular/platform-browser": "19.0.5",
+        "@angular/platform-browser-dynamic": "19.0.5",
+        "@angular/router": "19.0.5",
         "@aws-amplify/ui-angular": "^5.0.29",
         "@aws-sdk/client-dynamodb": "^3.682.0",
         "@ng-icons/core": "^29.5.1",
         "@ng-icons/material-icons": "^29.5.1",
         "@ngneat/until-destroy": "^10.0.0",
-        "@ngrx/component-store": "^18.0.1",
-        "@ngrx/effects": "^18.0.1",
-        "@ngrx/entity": "^18.0.1",
-        "@ngrx/router-store": "^18.0.1",
-        "@ngrx/store": "^18.0.1",
+        "@ngrx/effects": "^19.0.0",
+        "@ngrx/entity": "^19.0.0",
+        "@ngrx/router-store": "^19.0.0",
+        "@ngrx/store": "19.0.0",
         "@tailwindcss/line-clamp": "^0.4.4",
-        "angular-i18next": "^18.0.0",
-        "angularx-qrcode": "^18.0.2",
+        "angularx-qrcode": "^19.0.0",
         "aws-amplify": "^6.6.5",
         "aws-jwt-verify": "^4.0.1",
         "axios": "^1.6.0",
