@@ -1,21 +1,22 @@
 import {
     ChangeDetectionStrategy,
     Component,
-    Inject,
+    inject,
     OnInit
 } from '@angular/core';
 import { ProfileSettingsViewComponent } from '../views/profile-settings-view.component';
 import { FormGroup } from '@angular/forms';
 import { ProfileSettingsForm } from '../../types/profile-settings-form.types';
 import { createSettingsForm } from '../../utils/profile-settings-form.utils';
-import { SUPPORTED_LANGUAGES } from '@sanctumlab/fe/shared';
+import {
+    I18nTranslateService,
+    SUPPORTED_LANGUAGES
+} from '@sanctumlab/fe/shared';
 import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { I18NEXT_SERVICE, ITranslationService } from 'angular-i18next';
 
 @UntilDestroy()
 @Component({
     selector: 'ngx-profile-settings-container',
-    standalone: true,
     imports: [ProfileSettingsViewComponent],
     template: `<ngx-profile-settings-view
         [form]="form"
@@ -26,19 +27,16 @@ import { I18NEXT_SERVICE, ITranslationService } from 'angular-i18next';
 export class ProfileSettingsContainerComponent implements OnInit {
     protected form!: FormGroup<ProfileSettingsForm>;
     protected readonly languageOptions = SUPPORTED_LANGUAGES;
-
-    constructor(
-        @Inject(I18NEXT_SERVICE) private i18NextService: ITranslationService
-    ) {}
+    private readonly i18NTranslateService = inject(I18nTranslateService);
 
     ngOnInit(): void {
         this.form = createSettingsForm({
-            language: this.i18NextService.language
+            language: this.i18NTranslateService.language
         });
         this.form.controls.language.valueChanges
             .pipe(untilDestroyed(this))
             .subscribe(async language => {
-                await this.i18NextService.changeLanguage(language);
+                await this.i18NTranslateService.changeLanguage(language);
             });
     }
 }
