@@ -1,45 +1,30 @@
 import {
-    I18NEXT_SERVICE,
-    I18NextModule,
-    ITranslationService
-} from 'angular-i18next';
-import {
-    APP_INITIALIZER,
     EnvironmentProviders,
-    importProvidersFrom,
     isDevMode,
-    LOCALE_ID,
-    makeEnvironmentProviders
+    makeEnvironmentProviders,
+    provideAppInitializer
 } from '@angular/core';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import HttpApi from 'i18next-http-backend';
+import i18next from 'i18next';
 
-export const provideInternationalization = (): EnvironmentProviders =>
+const i18NextFactoryStorybook = async () => {
+    return i18next
+        .use(LanguageDetector)
+        .use(HttpApi)
+        .init({
+            supportedLngs: ['en'],
+            fallbackLng: 'en',
+            debug: isDevMode(),
+            returnEmptyString: false,
+            ns: [],
+            backend: {
+                loadPath: 'locales/{{lng}}.{{ns}}.json'
+            }
+        });
+};
+
+export const provideInternationalizationStorybook = (): EnvironmentProviders =>
     makeEnvironmentProviders([
-        importProvidersFrom(I18NextModule.forRoot()),
-        {
-            provide: APP_INITIALIZER,
-            useFactory: (i18next: ITranslationService) => () =>
-                i18next
-                    .use(HttpApi)
-                    .use(LanguageDetector)
-                    .init({
-                        supportedLngs: ['en'],
-                        fallbackLng: 'en',
-                        debug: isDevMode(),
-                        returnEmptyString: false,
-                        ns: [],
-                        backend: {
-                            loadPath: 'locales/{{lng}}.{{ns}}.json'
-                        }
-                    }),
-            deps: [I18NEXT_SERVICE],
-
-            multi: true
-        },
-        {
-            provide: LOCALE_ID,
-            deps: [I18NEXT_SERVICE],
-            useFactory: (i18next: ITranslationService) => i18next.language
-        }
+        provideAppInitializer(async () => await i18NextFactoryStorybook())
     ]);
