@@ -5,23 +5,22 @@ import {
     InputStateType
 } from '../../../../types/internal/input-internal.types';
 import { NgClass } from '@angular/common';
-import { I18NextModule } from 'angular-i18next';
 import { InputValidationComponent } from '../../../internal/input-validation.component';
+import { I18nPipe } from '../../../../pipes/i18n.pipe';
 
 @Component({
     selector: 'ngx-clib-textarea-input',
-    standalone: true,
     imports: [
         FormsModule,
         ReactiveFormsModule,
         NgClass,
-        I18NextModule,
-        InputValidationComponent
+        InputValidationComponent,
+        I18nPipe
     ],
     template: `<div>
         <label class="form-control w-full">
             <div class="label">
-                <span class="label-text">{{ label | i18nextEager }}</span>
+                <span class="label-text">{{ label | i18nTranslate }}</span>
             </div>
             <textarea
                 tabindex="0"
@@ -29,7 +28,7 @@ import { InputValidationComponent } from '../../../internal/input-validation.com
                 [attr.id]="id"
                 [attr.name]="id"
                 [formControl]="control"
-                [placeholder]="placeholder | i18nextEager"
+                [placeholder]="placeholder | i18nTranslate"
                 [attr.aria-label]="label"
                 [autofocus]="autofocus"
                 autocomplete="false"
