@@ -10,17 +10,16 @@ import { SidebarCategoryItem } from '../../../types/organisms/sidebar.types';
 import { LogoComponent } from '../../atoms/logo/logo.component';
 import { FormControl, ReactiveFormsModule } from '@angular/forms';
 import { IconButtonComponent } from '../../atoms/icon-button/icon-button.component';
-import { I18NextModule } from 'angular-i18next';
+import { I18nPipe } from '../../../pipes/i18n.pipe';
 
 @Component({
     selector: 'ngx-clib-sidebar',
-    standalone: true,
     imports: [
         NgIcon,
         LogoComponent,
         ReactiveFormsModule,
         IconButtonComponent,
-        I18NextModule
+        I18nPipe
     ],
     templateUrl: 'sidebar.component.html',
     styleUrls: [],
