@@ -1,10 +1,6 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { Component, importProvidersFrom, Input } from '@angular/core';
-import { I18NextModule } from 'angular-i18next';
-import {
-    MarkdownInputComponent,
-    provideMockInputValidationConfiguration
-} from '@sanctumlab/fe/component-library';
+import { Component, Input } from '@angular/core';
+import { provideMockInputValidationConfiguration } from '@sanctumlab/fe/component-library';
 import { MenuItemViewComponent } from './menu-item-view.component';
 import {
     AuthenticationService,
@@ -38,7 +34,6 @@ describe('MenuItemViewComponent', () => {
         TestBed.configureTestingModule({
             imports: [MenuItemViewComponent],
             providers: [
-                importProvidersFrom(I18NextModule.forRoot()),
                 provideMockInputValidationConfiguration(),
                 provideMockAuthConfiguration(),
                 {
