@@ -1,8 +1,7 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { importProvidersFrom } from '@angular/core';
-import { I18NextModule } from 'angular-i18next';
 import { provideMockInputValidationConfiguration } from '@sanctumlab/fe/component-library';
 import { ProfileSettingsPageComponent } from './profile-settings-page.component';
+import { I18nTranslateService } from '@sanctumlab/fe/shared';
 
 describe('ProfileSettingsPageComponent', () => {
     let component: ProfileSettingsPageComponent;
@@ -12,8 +11,11 @@ describe('ProfileSettingsPageComponent', () => {
         TestBed.configureTestingModule({
             imports: [ProfileSettingsPageComponent],
             providers: [
-                importProvidersFrom(I18NextModule.forRoot()),
-                provideMockInputValidationConfiguration()
+                provideMockInputValidationConfiguration(),
+                {
+                    provide: I18nTranslateService,
+                    useValue: { changeLanguage: jest.fn(), language: jest.fn() }
+                }
             ]
         }).compileComponents();
     }));
